refactor(db): Merge dbWrapperBase and rmWrapper to humongous dbWrapper
