+++ configmode.py
@@ -2,7 +2,7 @@ import os
 import sys
 from threading import Thread
 
-from db.dbWrapperBase import DbWrapperBase
+from db.DbWrapper import DbWrapper
 from db.DbFactory import DbFactory
 from utils.MappingManager import MappingManagerManager, MappingManager
 from utils.logging import initLogging, logger
@@ -24,7 +24,7 @@ def create_folder(folder):
         os.makedirs(folder)
 
 
-def start_madmin(args, db_wrapper: DbWrapperBase, ws_server, mapping_manager: MappingManager, data_manager, deviceUpdater, jobstatus):
+def start_madmin(args, db_wrapper: DbWrapper, ws_server, mapping_manager: MappingManager, data_manager, deviceUpdater, jobstatus):
     from madmin.madmin import madmin_start
     madmin_start(args, db_wrapper, ws_server, mapping_manager, data_manager, deviceUpdater, jobstatus)
