+++ route/RouteManagerLeveling.py
@@ -1,7 +1,7 @@
 import collections
 import time
 from typing import List
-from db.dbWrapperBase import DbWrapperBase
+from db.DbWrapper import DbWrapper
 from route.RouteManagerBase import RoutePoolEntry
 from route.RouteManagerQuests import RouteManagerQuests
 from utils.logging import logger
@@ -44,7 +44,7 @@ class RouteManagerLeveling(RouteManagerQuests):
             any_at_all = len(origin_local_list) > 0 or any_at_all
         return any_at_all
 
-    def __init__(self, db_wrapper: DbWrapperBase, dbm, area_id, coords: List[Location], max_radius: float,
+    def __init__(self, db_wrapper: DbWrapper, dbm, area_id, coords: List[Location], max_radius: float,
                  max_coords_within_radius: int, path_to_include_geofence: str, path_to_exclude_geofence: str,
                  routefile: str, mode=None, init: bool = False, name: str = "unknown", settings: dict = None,
                  level: bool = False, calctype: str = "optimized", joinqueue=None):
