+++ worker/WorkerQuests.py
@@ -5,7 +5,7 @@ from threading import Event, Thread
 from typing import List
 from datetime import datetime, timedelta
 
-from db.dbWrapperBase import DbWrapperBase
+from db.DbWrapper import DbWrapper
 from mitm_receiver.MitmMapper import MitmMapper
 from ocr.pogoWindows import PogoWindows
 from utils.MappingManager import MappingManager
@@ -45,7 +45,7 @@ class WorkerQuests(MITMBase):
         return ["pokestops"]
 
     def __init__(self, args, id, last_known_state, websocket_handler,  mapping_manager: MappingManager,
-                 routemanager_name: str, db_wrapper: DbWrapperBase, pogo_window_manager: PogoWindows, walker,
+                 routemanager_name: str, db_wrapper: DbWrapper, pogo_window_manager: PogoWindows, walker,
                  mitm_mapper: MitmMapper):
         MITMBase.__init__(self, args, id, last_known_state, websocket_handler,
                           mapping_manager=mapping_manager, routemanager_name=routemanager_name,
